Added the stopiferror function and provinces list.
Added the stopiferror function that stops running when the function call generates a one, which is the error return code when the SQL code sent through ssh_exec_wait generates an error. Also defied some objects for the Canadian provinces - a list of provinces and hash objects linking regions with the descriptive regional name and linking regions with the provinces they contain.